[MNY-79]: Reorgs tests; adds Comparable
